avoiding await on thr root level as it doesnt work on some older versions of node that I need to use

+++ install.sh
@@ -26,7 +26,14 @@ if command_exists ttls_forward_proxy_server; then
   fi
 fi
 
+
+../fbsource/xplat/third-party/node/bin/node install.mjs
+
 pushd $DIR &> /dev/null
-  npm install .
+  if command_exists yarn; then
+    yarn
+  elif command_exists npm; then
+    npm install .
+  fi
   ./install.mjs
 popd &> /dev/null
